refactored decision variables API to get it out of symbols
