+++ sysopt/symbolic/__init__.py
@@ -1,5 +1,5 @@
 """Sysopt symbolic and function manipulations."""
 
 from sysopt.symbolic.symbols import *
-from sysopt.symbolic.scalar_ops import *
-from sysopt.symbolic.decision_variables import PiecewiseConstantSignal
+from sysopt.symbolic.decision_variables import *
+from sysopt.symbolic.problem_data import *
