+++ sysopt/symbolic/symbols.py
@@ -1,5 +1,4 @@
 """Functions and factories to create symbolic variables."""
-import abc
 import weakref
 from abc import ABCMeta, abstractmethod
 from collections import defaultdict
@@ -18,22 +17,6 @@ SymbolicAtom = NewType('SymbolicAtom', Union['Variable', 'Parameter'])
 SymbolicArray = NewType(
     'SymbolicArray', Union[List[SymbolicAtom], Tuple[SymbolicAtom]]
 )
-
-
-def find_param_index_by_name(block, name: str):
-    try:
-        return block.find_by_name('parameters', name)
-    except (AttributeError, ValueError):
-        pass
-    try:
-        return block.parameters.index(name)
-    except ValueError:
-        pass
-    try:
-        return block.parameters.index(f'{str(block)}/{name}')
-    except ValueError:
-        pass
-    raise ValueError(f'Could not find parameter \'{name}\' in block {block}.')
 
 
 class Matrix(np.ndarray):
@@ -303,7 +286,6 @@ def power(base, exponent):
     return base ** exponent
 
 
-
 @register_op(string='+')
 def add(lhs, rhs):
     return lhs + rhs
@@ -313,6 +295,7 @@ def add(lhs, rhs):
 def sub(lhs, rhs):
     return lhs - rhs
 
+
 def is_scalar(item):
     if isinstance(item, (float, int, complex)):
         return True
@@ -321,6 +304,7 @@ def is_scalar(item):
     except AttributeError:
         pass
     return False
+
 
 @register_op(shape_func=matmul_shape, string='@')
 def matmul(lhs, rhs):
@@ -407,12 +391,13 @@ class PathInequality(Inequality):
             the constraint is violated.
 
         """
+
         c = Variable('c')
         rho = regulariser
         g = self.to_graph()
         # TODO: fix me
         # pylint: disable=import-outside-toplevel
-        from sysopt.symbolic.scalar_ops import exp
+        from sysopt.backends.scalar_ops import exp
 
         return c, exp(rho * (c - g)) / (alpha * rho)
 
@@ -575,9 +560,7 @@ def recursively_apply(graph: 'ExpressionGraph',
         return leaf_function(graph.value)
     if isinstance(graph, GraphWrapper):
         return recursively_apply(graph.graph, trunk_function, leaf_function)
-    if isinstance(graph, (Parameter, Variable)):
-        return leaf_function(graph)
-    if isinstance(graph, (int, float, np.ndarray)):
+    if not isinstance(graph, ExpressionGraph):
         return leaf_function(graph)
     sorted_nodes = graph.get_topological_sorted_indices()
     trunk_indices = {i for i in sorted_nodes if i in graph.edges}
@@ -897,118 +880,6 @@ numpy_handlers.update(
 )
 
 
-def symbolic_vector(name, length=1):
-    return Variable(name, shape=(length, ))
-
-
-class Variable(Algebraic):
-    """Symbolic type for a free variable."""
-    is_symbolic = True
-    __array_ufunc__ = None
-
-    def __init__(self, name=None, shape=scalar_shape):
-        self._shape = (shape, ) if isinstance(shape, int) else shape
-        self.name = name
-
-    def __str__(self):
-        shape_str = 'x'.join(str(n) for n in self.shape)
-        return f'{self.name}^({shape_str})'
-
-    def __repr__(self):
-        if self.name is not None:
-            return f'{self.__class__.__name__}({self.name}, {self.shape})'
-        else:
-            return 'unnamed_variable'
-
-    @property
-    def shape(self):
-        return self._shape
-
-    def symbols(self):
-        return {self}
-
-    def __hash__(self):
-        return hash(id(self))
-
-    def __cmp__(self, other):
-        return id(self) == id(other)
-
-
-_t = Variable('time')
-
-
-def resolve_parameter_uid(block, index):
-    name = block.parameters[index]
-    return hash(name)
-
-
-class Parameter(Algebraic):
-    """Symbolic type for variables bound to a block parameter.
-
-    Args:
-        block: The model block from which to derive the symbolic parameter.
-        parameter: Index or name of the desired symbolic parameter.
-
-    """
-    _table = {}
-
-    def __new__(cls, block, parameter: Union[str, int]):
-
-        if isinstance(parameter, str):
-            index = find_param_index_by_name(block, parameter)
-        else:
-            index = parameter
-        assert 0 <= index < len(block.parameters),\
-            f'Invalid parameter index for {block}: got {parameter},'\
-            f'expected a number between 0 and {len(block.parameters)}'
-
-        uid = resolve_parameter_uid(block, index)
-
-        try:
-            obj = Parameter._table[uid]
-            return obj
-        except KeyError:
-            pass
-        obj = Algebraic.__new__(cls)
-        obj.__init__()
-        setattr(obj, 'uid', uid)
-        setattr(obj, 'index', index)
-        setattr(obj, '_parent', weakref.ref(block))
-        Parameter._table[uid] = obj
-        return obj
-
-    def __hash__(self):
-        return hash(self.uid)
-
-    def __cmp__(self, other):
-        try:
-            return self.uid == other.uid
-        except AttributeError:
-            return False
-
-    def get_source_and_slice(self):
-        return self._parent(), slice(self.index, self.index + 1, None)
-
-    @property
-    def name(self):
-        parent = self._parent()
-        return parent.parameters[self.index]
-
-    def __repr__(self):
-        return self.name
-
-    @property
-    def shape(self):
-        return scalar_shape
-
-    def symbols(self):
-        return {self}
-
-    @staticmethod
-    def from_block(block):
-        return [Parameter(block, i) for i in range(len(block.parameters))]
-
-
 @register_op()
 def evaluate_signal(signal, t):
     return signal(t)
@@ -1243,6 +1114,45 @@ def as_vector(arg):
     )
 
 
+class Variable(Algebraic):
+    """Symbolic type for a free variable."""
+    is_symbolic = True
+    __array_ufunc__ = None
+
+    def __init__(self, name=None, shape=scalar_shape):
+        self._shape = (shape, ) if isinstance(shape, int) else shape
+        self._name = name
+
+    def __str__(self):
+        shape_str = 'x'.join(str(n) for n in self.shape)
+        return f'{self.name}^({shape_str})'
+
+    def __repr__(self):
+        if self.name is not None:
+            return f'{self.__class__.__name__}({self.name}, {self.shape})'
+        else:
+            return 'unnamed_variable'
+
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def shape(self):
+        return self._shape
+
+    def symbols(self):
+        return {self}
+
+    def __hash__(self):
+        return hash(id(self))
+
+    def __cmp__(self, other):
+        return id(self) == id(other)
+
+
+_t = Variable('time')
+
 def get_time_variable():
     return _t
 
@@ -1368,7 +1278,7 @@ def extract_quadratures(graph: Union[ExpressionGraph, Quadrature]) \
         -> Tuple[Algebraic, Variable, ExpressionGraph]:
 
     quadratures = {}
-    if isinstance(graph, (Variable, Parameter)):
+    if isinstance(graph, (Variable)):
         return graph, None, None
 
     if isinstance(graph, Quadrature):
@@ -1415,7 +1325,7 @@ def extract_quadratures(graph: Union[ExpressionGraph, Quadrature]) \
 
 class ConstantFunction(Algebraic):
     """Wrap a constant value and treat it like a function."""
-    def __init__(self, value, arguments: List[Union[Variable, Parameter]]):
+    def __init__(self, value, arguments: List[Variable]):
         if isinstance(value, np.ndarray):
             self.value = value.view(Matrix)
         elif isinstance(value, (list, tuple)):
